refactor: application層の修正
